_internal/viewer3d/requirements.txt
PyQt6>=6.0.0
PyOpenGL>=3.1.0
numpy>=1.24.0
PyOpenGL-accelerate>=3.1.0
pillow>=10.0.0
pyqtconfig>=0.9.1
